Homework Progress
Completed box and whisker plot with outliers. Something happened to my second pie graph when I ran this from the top, so next time I log in to work on it I need to look up the previous version of code and just copy and paste the cell that worked originally.